libmlx5: Added new poll_cq according to the new CQE format
When working with CQE  version 1, the CQE format is different, therefore
we need a new poll_cq function that is compatible with the new CQE
format.

When  allocating the user-context, we decide on the CQE version, and
according to this, we choose which poll_cq  function to use; the old one
that supports the old CQE format or the new one.

Issue: 564625

Change-Id: I8c3174b18ea8e5857fd3a02333f5a8495b25c878
Signed-off-by: Haggai Abramovsky <hagaya@mellanox.com>