libmlx5: Added QP and XSRQ create/destroy flow with user index
When wokring with CQE version 1, the library allocates a user-index for
each new QP/XSRQ, and this user-index is passed to the kernel.

Also in the destruction of a QP/XSRQ, the library needs to free the
user-index, so it can be reused.

In this stage, the library still doesn't work with CQE version 1,
therefore we prepared the user-index in the driver data,  but don't pass
it to the kernel.

Issue: 564625

Change-Id: Ifce2c2bb53536230bbbf54246d38f4b0c31de715
Signed-off-by: Haggai Abramovsky <hagaya@mellanox.com>