Add ibv_query_values support
In order to query the current HCA's core clock, libmlx5 should
support ibv_query_values verb. Querying the hardware's cycles
register is done by mmaping this register to user-space.
Therefore, when libmlx5 initializes we mmap the cycles register.
This assumes the machine's architecture places the PCI and memory in
the same address space.

Change-Id: Ibc31b26a75a4cf4ecc152f9228b34ec22634fa10
Signed-off-by: Matan Barak <matanb@mellanox.com>